artifacts/krimajlis/requirements.txt
flask
flask-cors
yfinance
numpy
pandas
requests
apscheduler
supabase
psycopg2-binary
gunicorn
python-dotenv
